+++ bin/install-dependencies.sh
@@ -22,11 +22,11 @@ installDependency flux-json-api https://github.com/fluxfw/flux-json-api/archive/
 
 installDependency flux-loading-api https://github.com/fluxfw/flux-loading-api/archive/refs/tags/v2022-11-04-1.tar.gz
 
-installDependency flux-localization-api https://github.com/fluxfw/flux-localization-api/archive/refs/tags/v2022-11-03-1.tar.gz
+installDependency flux-localization-api https://github.com/fluxfw/flux-localization-api/archive/refs/tags/v2022-11-07-1.tar.gz
 
 installDependency flux-pwa-api https://github.com/fluxfw/flux-pwa-api/archive/refs/tags/v2022-11-03-1.tar.gz
 
-installDependency flux-pwa-generator-api https://github.com/fluxfw/flux-pwa-generator-api/archive/refs/tags/v2022-11-02-1.tar.gz
+installDependency flux-pwa-generator-api https://github.com/fluxfw/flux-pwa-generator-api/archive/refs/tags/v2022-11-07-1.tar.gz
 
 installDependency flux-settings-api https://github.com/fluxfw/flux-settings-api/archive/refs/tags/v2022-11-03-1.tar.gz