+++ src/Adapter/IdentificationNumber/IdentificationNumberElement.mjs
@@ -8,6 +8,7 @@ import { TitleElement } from "../Title/TitleElement.mjs";
 /** @typedef {import("./confirmedIdentificationNumberFunction.mjs").confirmedIdentificationNumberFunction} confirmedIdentificationNumberFunction */
 /** @typedef {import("../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
 /** @typedef {import("./IdentificationNumber.mjs").IdentificationNumber} IdentificationNumber */
+/** @typedef {import("../../Service/Label/Port/LabelService.mjs").LabelService} LabelService */
 /** @typedef {import("../../Libs/flux-localization-api/src/Adapter/Api/LocalizationApi.mjs").LocalizationApi} LocalizationApi */
 
 const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
@@ -34,6 +35,10 @@ export class IdentificationNumberElement extends HTMLElement {
      */
     #identification_number;
     /**
+     * @type {LabelService}
+     */
+    #label_service;
+    /**
      * @type {LocalizationApi}
      */
     #localization_api;
@@ -44,15 +49,17 @@ export class IdentificationNumberElement extends HTMLElement {
 
     /**
      * @param {CssApi} css_api
+     * @param {LabelService} label_service
      * @param {LocalizationApi} localization_api
      * @param {IdentificationNumber} identification_number
      * @param {confirmedIdentificationNumberFunction} confirmed_identification_number_function
      * @param {backFunction | null} back_function
      * @returns {IdentificationNumberElement}
      */
-    static new(css_api, localization_api, identification_number, confirmed_identification_number_function, back_function = null) {
+    static new(css_api, label_service, localization_api, identification_number, confirmed_identification_number_function, back_function = null) {
         return new this(
             css_api,
+            label_service,
             localization_api,
             identification_number,
             confirmed_identification_number_function,
@@ -62,16 +69,18 @@ export class IdentificationNumberElement extends HTMLElement {
 
     /**
      * @param {CssApi} css_api
+     * @param {LabelService} label_service
      * @param {LocalizationApi} localization_api
      * @param {IdentificationNumber} identification_number
      * @param {confirmedIdentificationNumberFunction} confirmed_identification_number_function
      * @param {backFunction | null} back_function
      * @private
      */
-    constructor(css_api, localization_api, identification_number, confirmed_identification_number_function, back_function) {
+    constructor(css_api, label_service, localization_api, identification_number, confirmed_identification_number_function, back_function) {
         super();
 
         this.#css_api = css_api;
+        this.#label_service = label_service;
         this.#localization_api = localization_api;
         this.#identification_number = identification_number;
         this.#confirmed_identification_number_function = confirmed_identification_number_function;
@@ -102,29 +111,21 @@ export class IdentificationNumberElement extends HTMLElement {
             return;
         }
 
-        if (post_result["network-error"]) {
+        if (post_result["error-messages"] !== null) {
+            for (const error_message of post_result["error-messages"]) {
+                this.#shadow.prepend(this.#form_element.addInvalidMessage(
+                    await this.#label_service.getErrorMessageLabel(
+                        error_message
+                    )
+                ));
+            }
+        } else {
             this.#shadow.prepend(this.#form_element.addInvalidMessage(
                 this.#localization_api.translate(
-                    "Network error!"
+                    "Please check your data!"
                 )
             ));
-            return;
         }
-
-        if (post_result["server-error"]) {
-            this.#shadow.prepend(this.#form_element.addInvalidMessage(
-                this.#localization_api.translate(
-                    "Server error!"
-                )
-            ));
-            return;
-        }
-
-        this.#shadow.prepend(this.#form_element.addInvalidMessage(
-            this.#localization_api.translate(
-                "Please check your data!"
-            )
-        ));
     }
 
     /**