+++ src/Adapter/IdentificationNumber/confirmedIdentificationNumberFunction.mjs
@@ -1,6 +1,6 @@
 /** @typedef {import("./ConfirmedIdentificationNumber.mjs").ConfirmedIdentificationNumber} ConfirmedIdentificationNumber */
-/** @typedef {import("../Post/PostClientResult.mjs").PostClientResult} PostClientResult */
+/** @typedef {import("../Post/PostResult.mjs").PostResult} PostResult */
 
 /**
- * @typedef {(confirmed_identification_number: ConfirmedIdentificationNumber) => Promise<PostClientResult>} confirmedIdentificationNumberFunction
+ * @typedef {(confirmed_identification_number: ConfirmedIdentificationNumber) => Promise<PostResult>} confirmedIdentificationNumberFunction
  */