+++ src/Adapter/IntendedDegreeProgram/chosenIntendedDegreeProgramFunction.mjs
@@ -1,6 +1,6 @@
 /** @typedef {import("./ChosenIntendedDegreeProgram.mjs").ChosenIntendedDegreeProgram} ChosenIntendedDegreeProgram */
-/** @typedef {import("../Post/PostClientResult.mjs").PostClientResult} PostClientResult */
+/** @typedef {import("../Post/PostResult.mjs").PostResult} PostResult */
 
 /**
- * @typedef {(chosen_intended_degree_program: ChosenIntendedDegreeProgram) => Promise<PostClientResult>} chosenIntendedDegreeProgramFunction
+ * @typedef {(chosen_intended_degree_program: ChosenIntendedDegreeProgram) => Promise<PostResult>} chosenIntendedDegreeProgramFunction
  */