+++ src/Adapter/IntendedDegreeProgram2/chosenIntendedDegreeProgram2Function.mjs
@@ -1,6 +1,6 @@
 /** @typedef {import("./ChosenIntendedDegreeProgram2.mjs").ChosenIntendedDegreeProgram2} ChosenIntendedDegreeProgram2 */
-/** @typedef {import("../Post/PostClientResult.mjs").PostClientResult} PostClientResult */
+/** @typedef {import("../Post/PostResult.mjs").PostResult} PostResult */
 
 /**
- * @typedef {(chosen_intended_degree_program_2: ChosenIntendedDegreeProgram2) => Promise<PostClientResult>} chosenIntendedDegreeProgram2Function
+ * @typedef {(chosen_intended_degree_program_2: ChosenIntendedDegreeProgram2) => Promise<PostResult>} chosenIntendedDegreeProgram2Function
  */