+++ src/Adapter/Salutation/Salutation.mjs
@@ -1,3 +1,5 @@
+/** @typedef {import("../Label/Label.mjs").Label} Label */
+
 /**
- * @typedef {{id: string, label: {[key: string]: string}}} Salutation
+ * @typedef {{id: string, label: Label}} Salutation
  */