+++ src/Adapter/UniversityEntranceQualification/chosenUniversityEntranceQualificationFunction.mjs
@@ -1,6 +1,6 @@
 /** @typedef {import("./ChosenUniversityEntranceQualification.mjs").ChosenUniversityEntranceQualification} ChosenUniversityEntranceQualification */
-/** @typedef {import("../Post/PostClientResult.mjs").PostClientResult} PostClientResult */
+/** @typedef {import("../Post/PostResult.mjs").PostResult} PostResult */
 
 /**
- * @typedef {(chosen_university_entrance_qualification: ChosenUniversityEntranceQualification) => Promise<PostClientResult>} chosenUniversityEntranceQualificationFunction
+ * @typedef {(chosen_university_entrance_qualification: ChosenUniversityEntranceQualification) => Promise<PostResult>} chosenUniversityEntranceQualificationFunction
  */